# apis: don't run getConfigletByName for every configlet in get_configl…